Split files for management
