+++ functions/src/functions/verifyOtpCode.ts
@@ -1,0 +1,97 @@
+import * as admin from "firebase-admin";
+import { CollectionReference } from "firebase-admin/firestore";
+import * as functions from "firebase-functions";
+import { totp } from "otplib";
+import config from "../config";
+import { verifyEmail } from "../helper/helper";
+
+admin.initializeApp();
+const db = admin.firestore();
+
+/**
+ * Verify the Otp code the user provide
+ * If the code is correct, using the email we search for user that have this email as "link_email"
+ * Generate custom token with that user's uid and send back to client for authentication
+ *
+ * Params for the function
+ * @params  {email:string, code:string}
+ */
+export const verifyOtpCode = functions
+  .runWith({
+    // allows the function to use environment secret OTP_SECRET
+    secrets: ["OTP_SECRET"],
+    // no more than 20 instances of the function should be running at once.
+    // More on https://cloud.google.com/functions/docs/configuring/max-instances
+    maxInstances: 20,
+  })
+  .https.onCall(async (data) => {
+    if (!data.email) {
+      throw new functions.https.HttpsError(
+        "invalid-argument",
+        `The function must be called with the argument "email" containing the email the code is sent to.`
+      );
+    }
+    if (!data.code) {
+      throw new functions.https.HttpsError(
+        "invalid-argument",
+        `The function must be called with the argument "code" containing the code received.`
+      );
+    }
+
+    const email: string = data.email.trim();
+    if (!verifyEmail(email)) {
+      throw new functions.https.HttpsError(
+        "invalid-argument",
+        `The value ${email} doesn't seem to be a valid email.`
+      );
+    }
+    /**
+     * the secret is generated by concatenating the secret and user's provided email
+     * we do that as a "salting" to the secret
+     */
+    const secret = config.otpSecret + email;
+    const code = data.code.trim();
+
+    // check whether the token is valid or not
+    const isValid = totp.check(code, secret);
+
+    if (!process.env.FUNCTIONS_EMULATOR) {
+      if (!isValid) {
+        return {
+          isSuccess: false,
+          message: "The code is incorrect",
+          token: null,
+        };
+      }
+      const userRef = db.collection("users") as CollectionReference<{
+        link_email: string;
+      }>;
+      const matchedUsers = await userRef.where("link_email", "==", email).get();
+
+      if (matchedUsers.empty) {
+        return {
+          isSuccess: false,
+          message: "the email is not linked to any existing user",
+          token: null,
+        };
+      }
+      if (matchedUsers.size > 1) {
+        throw new functions.https.HttpsError(
+          "internal",
+          `The email ${email} is linked to multiple accounts. This shouldn't happen`
+        );
+      }
+      const user = matchedUsers.docs[0];
+      const authenticationToken = await admin.auth().createCustomToken(user.id);
+
+      return {
+        isSuccess: true,
+        message: "authentication success",
+        token: authenticationToken,
+      };
+    } else {
+      return {
+        message: "something is wrong here. Please contact support",
+      };
+    }
+  });
